Improve parsing of two-digit years.
This extends the behavior of ScalarYear when handling two digit years
(and indeed any year in the range 1-137).  Previously, the code treated
years in the range 38-68 as errors, 69-99 as 19xx, and the rest as 20xx.
This change uses the PointerType value in the options to change the
behavior.  If PointerType.PAST is passed, 20-68 is treated as
19xx (so you can parse birth dates for old people, for example).  If
PointerType.FUTURE is passed then 1-68 is treated as 20xx.
Using PointerType.NONE treats 38-68 as an error, as before.

Note that this is a change of behavior, because Options defaults to
use PointerType.FUTURE.  To get the old behavior back, you will need to
set PointerType.NONE explicitly.